Pkg restructure: util_data.py -> io/, structure/
