+++ aiida_jutools/__init__.py
@@ -22,12 +22,20 @@ For developers:
 - Place larger classes in separate modules, smaller functions in respective ``module/util.py``.
 - Make all available at module level via import in ``module/__init__.py``. See existing files as templates.
 - Also import each new module in the top-level ``__init__.py``.
+- Prefix non-user tools with ``_`` to keep user namespace clean and organized.
+- Prefix all imports inside modules with ``_`` to keep user namespace clean and organized. See existing modules
+  for conventions, esp. wrt AiiDA imports.
+- Add ``typing`` hints wherever possible and sensible. See existing modules for examples.
+- When manipulating nodes, implement with 'load or create' pattern: load nodes if already exist, otherwise create.
+  Provide a ``dry_run:bool=True`` and verbosity options (``verbosity:int``, ``verbose:bool``, or ``silent:bool``).
 """
 __version__ = "0.1.0-dev1"
 
 # module-level imports
 import code
 import computer
+import io
 import logging
 import process_functions
+import structure
